Add TypeExpr AST accessors and consolidate type lowering
- Add is_union() and parts() methods to TypeExpr in ast.rs
- Simplify TypeRef::from_ast() to use new accessors
- Remove duplicate lower_type_ref() from lib.rs and signature.rs
- Fix: primitives now recognized inside type constructors (int? → Optional(Int))
